requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.1.2
djangorestframework==3.14.0
kombu==5.3.7
prompt_toolkit==3.0.47
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.1
vine==5.1.0
wcwidth==0.2.13
